Added new classes to improve OOP requirements
